+++ modules/core/src/main/java/net/ashald/envfile/EnvVarsProviderFactory.java
@@ -1,5 +1,7 @@
 package net.ashald.envfile;
 
+import java.util.Map;
+import java.util.function.Consumer;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -8,10 +10,14 @@ import java.util.function.Predicate;
 
 public interface EnvVarsProviderFactory {
 
-    @NotNull
-    EnvVarsProvider createProvider(boolean shouldSubstituteEnvVar);
+    /**
+     * @param baseEnvVars env vars defined in the run configuration
+     */
+    EnvVarsProvider createProvider(Map<String, String> baseEnvVars, Consumer<String> logger);
 
-    @NotNull String getTitle();
+    String getTitle();
+
+    boolean isEditable();
 
     @Nullable Predicate<String> getFileNamePredicate();