+++ modules/core/src/main/java/net/ashald/envfile/providers/runconfig/RunConfigEnvVarsProvider.java
@@ -1,33 +1,36 @@
 package net.ashald.envfile.providers.runconfig;
 
-import net.ashald.envfile.AbstractEnvVarsProvider;
-import org.jetbrains.annotations.NotNull;
+import lombok.AllArgsConstructor;
+import net.ashald.envfile.EnvVarsProvider;
+import net.ashald.envfile.exceptions.InvalidEnvFileException;
 
 import java.io.File;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
-public class RunConfigEnvVarsProvider extends AbstractEnvVarsProvider {
+@AllArgsConstructor
+public class RunConfigEnvVarsProvider implements EnvVarsProvider {
     public static final String PARSER_ID = "runconfig";
 
-    public RunConfigEnvVarsProvider(boolean shouldSubstituteEnvVar) {
-        super(shouldSubstituteEnvVar);
-    }
-
-    @NotNull
-    @Override
-    protected Map<String, String> getEnvVars(@NotNull Map<String, String> runConfigEnv, InputStream content, String path) {
-        return new HashMap<>(runConfigEnv);
-    }
+    private final Map<String, String> baseEnvVars;
 
     @Override
-    public boolean isEditable() {
-        return false;
-    }
+    public Map<String, String> getEnvVars(File file, boolean isExecutable, Map<String, String> context)
+            throws InvalidEnvFileException {
+        if (file != null) {
+            throw InvalidEnvFileException.format(
+                    "%s is not supposed to be configured with a file",
+                    this.getClass().getSimpleName()
+            );
+        }
 
-    @Override
-    public boolean isFileLocationValid(File file) {
-        return true; // no file needed
+        if (isExecutable) {
+            throw InvalidEnvFileException.format(
+                    "%s does not support executable mode",
+                    this.getClass().getSimpleName()
+            );
+        }
+
+        return new HashMap<>(baseEnvVars);
     }
 }