+++ modules/core/src/main/java/net/ashald/envfile/providers/runconfig/RunConfigEnvVarsProviderFactory.java
@@ -1,6 +1,5 @@
 package net.ashald.envfile.providers.runconfig;
 
-
 import net.ashald.envfile.EnvVarsProvider;
 import net.ashald.envfile.EnvVarsProviderFactory;
 import org.jetbrains.annotations.NotNull;
@@ -8,18 +7,24 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
+import java.util.Map;
+import java.util.function.Consumer;
+
 public class RunConfigEnvVarsProviderFactory implements EnvVarsProviderFactory {
 
-
-    @NotNull
     @Override
-    public EnvVarsProvider createProvider(boolean shouldSubstituteEnvVar) {
-        return new RunConfigEnvVarsProvider(shouldSubstituteEnvVar);
+    public EnvVarsProvider createProvider(Map<String, String> baseEnvVars, Consumer<String> logger) {
+        return new RunConfigEnvVarsProvider(baseEnvVars);
     }
 
     @Override
     public @NotNull String getTitle() {
         return "Run Config";
+    }
+
+    @Override
+    public boolean isEditable() {
+        return false;
     }
 
     @Override