+++ modules/core/src/main/java/net/ashald/envfile/providers/yaml/YamlEnvFileParserFactory.java
@@ -2,22 +2,41 @@ package net.ashald.envfile.providers.yaml;
 
 import net.ashald.envfile.EnvVarsProvider;
 import net.ashald.envfile.EnvVarsProviderFactory;
+import net.ashald.envfile.providers.EnvFileExecutor;
+import net.ashald.envfile.providers.EnvFileReader;
+import net.ashald.envfile.providers.SingleFileEnvVarsProvider;
 import org.jetbrains.annotations.NotNull;
+import org.yaml.snakeyaml.Yaml;
+
+import java.util.Map;
+import java.util.function.Consumer;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
 public class YamlEnvFileParserFactory implements EnvVarsProviderFactory {
+    private static final Yaml YAML = new Yaml();
 
     @Override
-    public @NotNull
-    EnvVarsProvider createProvider(boolean shouldSubstituteEnvVar) {
-        return new YamlEnvFileParser(shouldSubstituteEnvVar);
+    public EnvVarsProvider createProvider(Map<String, String> baseEnvVars, Consumer<String> logger) {
+        return SingleFileEnvVarsProvider.builder()
+                .reader(EnvFileReader.DEFAULT)
+                .executor(EnvFileExecutor.DEFAULT)
+                .parser(
+                        new YamlEnvFileParser(YAML)
+                )
+                .logger(logger)
+                .build();
     }
 
     @NotNull
     public String getTitle() {
         return "JSON/YAML";
+    }
+
+    @Override
+    public boolean isEditable() {
+        return true;
     }
 
     @Override