+++ modules/platform/src/main/java/net/ashald/envfile/platform/ui/table/EnvFilePathColumnInfo.java
@@ -1,10 +1,18 @@
 package net.ashald.envfile.platform.ui.table;
 
+import com.intellij.openapi.project.Project;
 import com.intellij.ui.JBColor;
 import com.intellij.ui.components.JBTextField;
 import com.intellij.util.ui.ColumnInfo;
 import com.intellij.util.ui.UIUtil;
+
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.val;
+import net.ashald.envfile.EnvVarsProviderFactory;
 import net.ashald.envfile.platform.EnvFileEntry;
+import net.ashald.envfile.platform.EnvVarsProviderExtension;
+import net.ashald.envfile.platform.ProjectFileResolver;
 import net.ashald.envfile.providers.runconfig.RunConfigEnvVarsProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -15,15 +23,26 @@ import javax.swing.table.DefaultTableCellRenderer;
 import javax.swing.table.TableCellEditor;
 import javax.swing.table.TableCellRenderer;
 import java.awt.Component;
+import java.io.File;
+
 
 public class EnvFilePathColumnInfo extends ColumnInfo<EnvFileEntry, String> {
-    public EnvFilePathColumnInfo() {
+
+    @NonNull
+    private final Project project;
+
+    @Builder
+    public EnvFilePathColumnInfo(@NotNull final Project project) {
         super("Path");
+        this.project = project;
     }
 
     @Override
     public boolean isCellEditable(EnvFileEntry envFileEntry) {
-        return envFileEntry.isEditable() && envFileEntry.getEnabled();
+        return EnvVarsProviderExtension.getParserFactoryById(envFileEntry.getParserId())
+                .map(EnvVarsProviderFactory::isEditable)
+                .map(value -> value && envFileEntry.isEnabled())
+                .orElse(false);
     }
 
     @Nullable
@@ -44,7 +63,7 @@ public class EnvFilePathColumnInfo extends ColumnInfo<EnvFileEntry, String> {
     }
 
     @Override
-    public TableCellRenderer getRenderer(final EnvFileEntry p0) {
+    public TableCellRenderer getRenderer(final EnvFileEntry entry) {
         return new DefaultTableCellRenderer() {
             @NotNull
             @Override
@@ -58,14 +77,17 @@ public class EnvFilePathColumnInfo extends ColumnInfo<EnvFileEntry, String> {
             ) {
                 final Component renderer = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
                 setBorder(null);
-                if (p0.getParserId().equals(RunConfigEnvVarsProvider.PARSER_ID)) { // TODO make generic
+                if (entry.getParserId().equals(RunConfigEnvVarsProvider.PARSER_ID)) { // TODO make generic
                     setText("<Run Configuration Env Vars>");
                     setForeground(UIUtil.getLabelDisabledForeground());
                 } else {
-                    setText(p0.getPath());
+                    setText(entry.getPath());
 
-                    if (p0.isEnabled()) {
-                        if (!p0.validatePath()) {
+                    if (entry.isEnabled()) {
+                        val validPath = ProjectFileResolver.DEFAULT.resolvePath(project, entry.getPath())
+                                .map(File::exists)
+                                .orElse(true);
+                        if (!validPath) {
                             setForeground(JBColor.RED);
                             setToolTipText("File doesn't exist!");
                         }